refactor(ui): fix all circular dependencies
