+++ frontend/src/components/recording/RecordingCardDaily.tsx
@@ -17,10 +17,11 @@ import LazyLoad from "react-lazyload";
 import { Link } from "react-router-dom";
 
 import MutationIconButton from "components/buttons/MutationIconButton";
+import { getVideoElement } from "components/player/utils";
 import VideoPlayerPlaceholder from "components/player/videoplayer/VideoPlayerPlaceholder";
 import { useAuthContext } from "context/AuthContext";
 import { useDeleteRecording } from "lib/api/recordings";
-import { getVideoElement, objHasValues } from "lib/helpers";
+import { objHasValues } from "lib/helpers";
 import {
   getDateStringFromDayjs,
   getDayjsFromDateString,
