+++ frontend/src/lib/api/cameras.ts
@@ -1,7 +1,8 @@
 import { UseQueryOptions, useQuery } from "@tanstack/react-query";
 import { useMemo } from "react";
 
-import { useInvalidateQueryOnEvent, viseronAPI } from "lib/api/client";
+import { viseronAPI } from "lib/api/client";
+import { useInvalidateQueryOnEvent } from "lib/api/utils";
 import * as types from "lib/types";
 
 type CamerasVariables = {
